feat: Three days forecast, UiUtils also changed models, fixed uiState and more

+++ app/src/main/java/ru/lonelywh1te/justweather/domain/models/CurrentWeather.kt
@@ -4,12 +4,12 @@ import java.util.Date
 
 data class CurrentWeather(
     val lastUpdated: Date,
-    val tempC: Double,
-    val tempF: Double,
+    val tempC: Int,
+    val tempF: Int,
     val condition: Condition,
     val windKph: Double,
     val windMph: Double,
-    val feelsLikeC: Double,
-    val feelsLikeF: Double,
+    val feelsLikeC: Int,
+    val feelsLikeF: Int,
     val uv: Double,
 )
