+++ app/src/main/java/ru/lonelywh1te/justweather/domain/models/Hour.kt
@@ -4,13 +4,13 @@ import java.util.Date
 
 data class Hour(
     val time: Date,
-    val tempC: Double,
-    val tempF: Double,
+    val tempC: Int,
+    val tempF: Int,
     val isDay: Boolean,
     val condition: Condition,
     val windKph: Double,
     val windMph: Double,
-    val feelsLikeC: Double,
-    val feelsLikeF: Double,
+    val feelsLikeC: Int,
+    val feelsLikeF: Int,
     val uv: Double,
 )
